+++ src/Core/Output.lua
@@ -40,6 +40,69 @@ local Output = {}
     end
 
     --[[--
+    Dumps the values of variables and tables in the output, then dies.
+
+    The dd() stands for "dump and die" and it's a helper function inspired by a PHP framework
+    called Laravel. It's used to dump the values of variables and tables in the output and stop
+    the execution of the script. It's only used for debugging purposes and should never be used
+    in an addon that will be released.
+
+    Given that it can't use the Output:out() method, there's no test coverage for dd(). After
+    all it's a test and debugging helper resource.
+
+    @param ... The variables and tables to be dumped
+
+    @usage
+        dd(someVariable)
+        dd({ key = 'value' })
+        dd(someVariable, { key = 'value' })
+    ]]
+    function Output:dd(...)
+        -- @TODO: Replace this once the Environment class is implemented <2024.04.21>
+        local inGame = os == nil
+
+        if not inGame then print('\n\n\27[32m-dd-\n') end
+
+        local function printTable(t, indent, printedTables)
+            indent = indent or 0
+            printedTables = printedTables or {}
+            local indentStr = string.rep(" ", indent)
+            for k, v in pairs(t) do
+                if type(v) == "table" then
+                    if not printedTables[v] then
+                        printedTables[v] = true
+                        print(indentStr .. k .. " => {")
+                        printTable(v, indent + 4, printedTables)
+                        print(indentStr .. "}")
+                    else
+                        print(indentStr .. k .. " => [circular reference]")
+                    end
+                else
+                    print(indentStr .. k .. " => " .. tostring(v))
+                end
+            end
+        end
+
+        for i, v in ipairs({...}) do
+            if type(v) == "table" then
+                print("[" .. i .. "] => {")
+                printTable(v, 4, {})
+                print("}")
+            else
+                print("[" .. i .. "] => " .. tostring(v))
+            end
+        end
+
+        -- this prevents os.exit() being called inside the game and also allows
+        -- dd() to be tested
+        if inGame then return end
+        
+        print('\n-end of dd-')
+        lu.unregisterCurrentSuite()
+        os.exit(1)
+    end
+
+    --[[--
     Formats a standard message with the addon name to be printed.
 
     @tparam string message The message to be formatted
@@ -122,6 +185,7 @@ local Output = {}
 
 -- sets the unique library output instance
 self.output = Output.__construct()
+self.dd = self.output.dd
 
 -- allows Output to be instantiated, very useful for testing
 self:addClass('Output', Output)